AYS-246 | Exceptions in `admin_user.util.exception` Have Been Moved to `user.util.exception` Package
