+++ src/test/java/org/ays/admin_user/service/impl/AdminRegistrationApplicationServiceImplTest.java
@@ -13,9 +13,6 @@ import org.ays.admin_user.model.entity.AdminRegistrationApplicationEntity;
 import org.ays.admin_user.model.enums.AdminRegistrationApplicationStatus;
 import org.ays.admin_user.model.mapper.AdminRegistrationApplicationEntityToAdminRegistrationApplicationMapper;
 import org.ays.admin_user.repository.AdminRegistrationApplicationRepository;
-import org.ays.admin_user.util.exception.AysAdminRegistrationApplicationNotExistByIdException;
-import org.ays.admin_user.util.exception.AysAdminRegistrationApplicationNotExistByIdOrStatusNotWaitingException;
-import org.ays.admin_user.util.exception.AysAdminRegistrationApplicationSummaryNotExistByIdException;
 import org.ays.common.model.AysPage;
 import org.ays.common.model.AysPageBuilder;
 import org.ays.common.util.AysRandomUtil;
@@ -27,6 +24,9 @@ import org.ays.user.model.entity.UserEntityV2;
 import org.ays.user.model.entity.UserEntityV2Builder;
 import org.ays.user.model.enums.UserStatus;
 import org.ays.user.repository.UserRepositoryV2;
+import org.ays.user.util.exception.AysAdminRegistrationApplicationNotExistByIdException;
+import org.ays.user.util.exception.AysAdminRegistrationApplicationNotExistByIdOrStatusNotWaitingException;
+import org.ays.user.util.exception.AysAdminRegistrationApplicationSummaryNotExistByIdException;
 import org.junit.jupiter.api.Assertions;
 import org.junit.jupiter.api.Test;
 import org.mockito.InjectMocks;
