+++ src/test/java/org/ays/admin_user/service/impl/AdminRegistrationCompleteServiceImplTest.java
@@ -7,9 +7,6 @@ import org.ays.admin_user.model.entity.AdminRegisterApplicationEntityBuilder;
 import org.ays.admin_user.model.entity.AdminRegistrationApplicationEntity;
 import org.ays.admin_user.model.enums.AdminRegistrationApplicationStatus;
 import org.ays.admin_user.repository.AdminRegistrationApplicationRepository;
-import org.ays.admin_user.util.exception.AysAdminRegistrationApplicationNotExistByIdOrStatusNotWaitingException;
-import org.ays.admin_user.util.exception.AysUserAlreadyExistsByEmailException;
-import org.ays.admin_user.util.exception.AysUserAlreadyExistsByPhoneNumberException;
 import org.ays.common.model.dto.request.AysPhoneNumberRequest;
 import org.ays.common.model.dto.request.AysPhoneNumberRequestBuilder;
 import org.ays.common.util.AysRandomUtil;
@@ -23,6 +20,9 @@ import org.ays.user.repository.PermissionRepository;
 import org.ays.user.repository.RoleRepository;
 import org.ays.user.repository.UserPasswordRepository;
 import org.ays.user.repository.UserRepositoryV2;
+import org.ays.user.util.exception.AysAdminRegistrationApplicationNotExistByIdOrStatusNotWaitingException;
+import org.ays.user.util.exception.AysUserAlreadyExistsByEmailException;
+import org.ays.user.util.exception.AysUserAlreadyExistsByPhoneNumberException;
 import org.ays.util.AysValidTestData;
 import org.junit.jupiter.api.Assertions;
 import org.junit.jupiter.api.Test;
